catch OSError instead of FileNotFoundError to resolve #15399
While finding available movie writers, if the executable is presented but
cannot be run, the import process will fail upon calling '_get_executable_info'
since it only handled CalledProcessError and FileNotFoundError.

In this commit, we catch OSError instead of FileNotFoundError in '__init__.py'
and log the error message in 'animation.py' to improve the import robustness.