Reuse png metadata handling of imsave() in FigureCanvasAgg.print_png().
This avoids duplicating the conversion of metadata to PngInfo and
revealed a bug in the priority between `metadata` and `pil_kwargs` in
imsave().

Note that because `np.asarray(self.buffer_rgba())` is already a RGBA
uint8 array, there is no colormapping step happening in imsave().

Ideally mplcairo should also be able to use imsave() for saving to png.