Deprecate Locator.refresh and associated helpers.
`Locator.refresh()` was called at certain places to let locators update
their internal state, typically based on the axis limits.  Instead, they
can just always consult the axis limits when called, if needed.  (This
should be more robust -- otherwise, a call to a locator basically always
needs to be preceded to a call to `refresh()` to ensure getting
up-to-date results.)

The deprecation is made slightly complex by the fact that we need to
check whether the user is using a third-party locator which overrides
`refresh()`, and warn in that case.

The associated helper methods `NavigationToolbar2.draw()` and
`ToolViewsPositions.refresh_locators()` (which called `refresh()` and
then `canvas.draw_idle()` are deprecated, and should be replaced by
calls to `draw_idle()` on the corresponding canvas once the deprecation
period ends.