Consistently use realpaths to build XObject names
Before this comment, the PDF backend used two inconsistent ways to build the
name for XObjects required for Unicode glyphs above code point 255. The first
method is based on the file name of the font returned by
`font_manager.findfont()`. This method is used to refer to the XObject glyphs
when drawing text. When the XObjects are created inside `writeFonts()` the
realpath of the font returned by `cbook.get_realpath_and_stat()` is used
instead. The realpath differs from the file name of the font if the font is a
symbolic link. In this case, the produced PDF contains invalid references to
XObjects.

This commit adds a safe-guard in `PdfFile._get_xobject_symbol_name()`. The
returned XObject names are always based on the realpath, even if the supplied
path is a link.  This achieves consistent XObject names and prevents invalid PDF
files.