Fix stepfilled histogram polygon bottom perimeter
The polygon generated by hist() when histtype='stepfilled' and bottom is not
None contained an invalid starting point with a y-value of zero resulting in a
diagonal line at the bottom of the first bin. This commit fixes the computation
of the starting point and thus removes the drawing artifacts from the first bin.