Closes #15796: latex_elements defined twice in doc/conf.py
This commit reactivates original latex_elements['preamble'] key which
was shadowed by the bug about latex_elements being defined twice in the
conf.py.

It removes from this latex preamble the loading of amsmath, amssymb,
amsfonts as this is already added by Sphinx, and of txfonts because this
is not compatible with fontspec and the already executed command
\setmainfont{DejaVu Serif}.

Memo: expdlist package breaks PDF build if a description list ends up in
a table cell, as happens actually with Matplotlib.tex file. But, by
luck, the enumitem LaTeX package modifications of LaTeX description
lists masks the expdlist potential breakage.