Remove some references to Py2.
The checks in `conf.py` and `matplotlib/__init__.py` are not useful
because the files are in fact not Py2-compatible (this has been the case
for `__init__.py` for a while...) due to f-strings/nonlocal, so someone
trying to run the file on Py2 will get a SyntaxError instead of the
error message.

In image.py the cast to int remains necessary even on Py3 because
round(np.float) returns a np.float, not a np.int.