Colorbar cleanup.
Deprecate on_mappable_changed in favor of update_normal (they're the
same now (except for a logging call) now that colorbars just use the
norm of the mappable).

Deprecate update_bruteforce in favor of update_normal -- it's not used
anywhere except in axes_grid, but that's just because whoever introduced
update_normal forgot to update axes_grid at the same time.

axes_grid.colorbar is already deprecated but until its complete removal,
we need to backport update_normal to it...