Deprecate update_datalim_bounds.
This method was introduced in 56d43e6 with a note that it would be more
efficient to use it in some cases, then the note went away in ad74c3f
(2008) without the method ever being used either before or after.  In
any case, it is simple enough to be reimplemented (possibly inline) if
we really need it.