Deprecate MovieWriter.{exec,args}_key, making them private.
Builtin writer subclasses additionally define the _exec_key and
_args_key attributes, which indicate the rcParams entries where the
path to the executable and additional command-line arguments to the
executable are stored.  Third-party writers cannot meaningfully set
these as they cannot extend rcParams with new keys.

Hence they are really an internal way of mapping builtin writers to
rcParams and should not be exposed.