Deprecate Artist.{set,get}_contains.
I don't think it makes sense to allow overriding contains() checks
per instance -- and if someone really need that, they can use a custom
artist subclass or even just monkeypatch the contains method; if they
need it for pick events there's still Artist.set_picker which is not
going away.

This came in in 8111a3a (2007) and was apparently never used or tested.