Fix interaction with unpickled 3d plots.
In order to reset the mouse interaction callbacks on unpickled 3d plots,
the callback registry really needs to be on the Figure object rather
than the Canvas, because the canvas doesn't exist yet when the 3d axes
is being unpickled (it is only set on the figure at the very end of
unpickling).

So move the callback registry to the figure (with a proxy property on
the canvas).

Then, add a private mechanism to pickle select callbacks, and combine
everything together.

Test with e.g.
```
import matplotlib.pyplot as plt
import pickle

fig = plt.figure()
fig.add_subplot(111, projection='3d')

p = pickle.dumps(fig)
plt.close("all")

pickle.loads(p)

plt.show()
```