CI: Remove PYTHONUNBUFFERED=1 on Appveyor
It is there for unknown reason, since Appveyor setup. I do not expect something
to break, probably it was somehow related to multiprocess bug in Python 3 prior
to 3.6.3, and we have a fix for it already.

I expect to see some insignificant speed-up (should be much more on pdf tests).