Dedupe ImageGrid doc from tutorial and docstring.
ImageGrid is so little-maintained that I don't think we should have
duplicate argument lists in the tutorial and the docstring, given that
they *will* go out of sync (for more commonly used APIs, sure, we can
maintain the duplication, but here it's not worth it).  Add to the
docstring the relevant parts of the tutorial, and make the tutorial
point to the API docs.