MNT: rename internal variable
When I wrote this code ~4.5 years ago I was a bit cranky and used a
very sarcastic variable name (snowflake) to handle a class of special
cases when parsing math text.  It has recently come to light that this
name can make it into user visible error messages.  This renames the
variable to a less sarcastic and more descriptive variable name (it
handles the case of symbols who's names start with a valid accent; we
should parse it as the symbol, not an accent on the rest of the symbol
name).