Rework pylab docstring.
The old pylab module docstring was not rendered anywhere in the html
docs (it was only visible with pydoc or with IPython's `pylab?`), and is
outdated (e.g. it doesn't even list viridis as colormap).

Instead of trying to list the functions provided, replace the docstring
by the warning discouraging the use of pylab, and use that as entry in
the html docs.