Delay checking for existence of postscript distillers.
Currently, if one has `ps.usedistiller` set in their matplotlibrc,
matplotlib will immediately check for the presence of ghostscript (and
possibly pstopdf) at import time and run them in subprocesses to check
their version, even though they may not be needed at all (if the user is
not going to save a postscript image); this can quite slow down import
time.  Compare e.g. with `text.usetex` which requires the presence of
a `tex` install, but doesn't check for its presence until it is actually
needed.

Delay the checking for the presence of the distillers until we actually
need them.  For consistency with the old behavior, don't error out if
they are missing, but just emit a warning.