Fix _preprocess_data for Py3.9.
On Py3.9 BoundArguments becomes a plain dict, so move_to_end doesn't
exist anymore and the whole thing crashes with an AttributeError.
(Travis currently fails the Py3.9 run.)

As it turns out there's a simpler implementation of the whole thing
anyways, one can just take whatever args/kwargs were explicitly passed
in and, if "label" needs to be added, always add it to kwargs.