nbagg: Don't send events if manager is disconnected.
This occurs, e.g., when saving figures, as one of the first things
`print_figure` does is to remove the figure manager. This does mean that
nbagg won't be able to set the wait cursor when saving a file, but
before 0944dcbe6dd755fd51a91f9ca2eb57c4ae798624 that was not being done,
and after, saving was broken, so there's no net loss.

Fixes #16721.