Deprecate NavigationToolbar2QT.parent.
This attribute shadows the normal Qt method parent() (which also returns
the parent window...), which is a source of confusion.  After the
deprecation period passes, the parent property can be deleted and
parent() will simply be the method inherited from QToolBar.

Note that the attribute is not present on other backend toolbars either.