update units.ConversionInterface to have `to_numeric` and `from_numeric`
This change updates `matplotlib.units.ConversionInterface` to have a
`to_numeric` instead of a `convert` method, and adds a `from_numeric`
method. The method `convert` is renamed to `to_numeric` and similarly
named methods are updated accordingly. The renaming is done to make
`convert` less ambiguous with a forward and backward conversion on the
`ConversionInterface`.

Adding a `from_numeric` method standardizes the way to convert back to a
native object from an internal numeric value. This allows registered
converters to provide more specific conversions which will allow
formatters and locators to provide an implementation which is not as
tied to the internal values Matplotlib uses.