Fix use of psfrags in ps backend + usetex.
Looks like psfrags interacts badly with `\newgeometry`, so we need to
use `\geometry` instead in the custom preamble to set the paper size.

In turn this means that `\usepackage{geometry}` needs to move to before
the custom preamble into `TexManager._get_preamble()`, and thus we
can't pass the dvips option to it anymore, but fortunately dvips is
documented as the default driver anyways (section 5.6 of `geometry`
docs).  Also, in order to keep everything under 79 characters wide,
change `{72in,72in}` to just `72in` (which is synonymous per section
4.3.3 of `geometry` docs).

Also insert some extra newlines in the generated tex file for ease of
debugging.