Add test for close_event.
Test that a single close_event is triggered when a window is closed.

This caught a bug in the gtk3 backends, which used to emit *two*
close_events (see comment there; a better fix would be welcome); and
another in the wx backends (which emit *no* close_events when the window
is closed by pressing "q" -- now fixed).