Fix DPI scaling on macosx
Rename _device_scale to _dpi_ratio, as expected by
Figure.set_size_inches(). Also, make sure the value of _dpi_ratio on
instances of FigureCanvasMac are updated *before* setting figure DPI so
that the proper value is available for downstream code. This is part of
the fix needed to make set_size_inches work on macosx.