Change the description of Rectangle's xy parameter
matplotlib.patches.Rectangle's first parameter, xy, is previously only documented for data plotting functions in which xy means left and bottom coordinates. However, in imshow(), the vertical axis's direction is inverted, and xy means left and top coordinates. See #15401.

In this PR, a fix is made to explain xy as starting coordinates so as to make it suitable for both cases
Closes #15433
Taken things from #15433.