Fix missing-references.json.
I think what happened is that the way nested classes are referred to in
inheritance diagrams has improved (the qualified name is more correct
now), which caused missing-references.json to go out of sync.
Unfortunately this means missing-references.json can either be
compatible with sphinx<3.0 or with sphinx>3.0, but to be compatible with
both we'd likely need to generate one with both versions and merge the
two json files, which seems not worth it.  Instead, just drop nitpicky
mode for older sphixes (sphinges?).