Don't create a statusbar in Qt, wx backends.
Display the coordinates text on the right of the toolbar, consistently
with the GTK3, wx, and (I think?) macosx backends.  This helps when
embedding Matplotlib in larger GUIs, because Matplotlib may not
necessarily control the statusbar (which is typically global for the
whole window) but controls the toolbar if there's one.

If we decide to go this route we could probably also kill the status bar
for toolmanager.