Deprecate ScalarMappable.check_update and associated machinery.
It's a rather complex machinery that's ultimately just used by
Collection to know whether the alpha array has been updated so that the
value-to-rgba mapping needs to be updated.  Users can't connect to it to
know whether the mapped values have been updated, and they can't even
use it to check whether the *alpha* values have been updated because
this unsets the marker and thus would mess up Collection's own use of
it.  For now, just deprecate its public use to ultimately make it
private and perhaps replace it later by something less brittle (e.g.,
normal a normal callback registry).