Fix handling single contour level out of data range
Previously, the behavior of ContourSet was special when a single contour
level was specified that was out of range of the given z array.

In that case, instead of the given value the minimum of the dataset was
contoured, instead. This is very unexpected, in my opinion.

This simple change simply makes no contour line at all, instead.