Make all parameters of ColorbarBase, except `ax`, keyword-only.
This is consistent with Colorbar.  Moreover, this will make it possibly
to ultimately merge Colorbar into ColorbarBase -- currently, they only
differ by the fact that a ColorbarBase doesn't have an associated
ScalarMappable, but we can just construct an empty ScalarMappable in
that case (this is already what we document as recommended approach to
construct colorbars not attached to an artist --
`fig.colorbar(ScalarMappable(norm=..., cmap=...))`.  Right now this
merging is *nearly* possible except for the fact that the constructors'
signatures are a pain to combine.