Simplify some axisartist code.
In a few places, axisartist temporarily sets the transform of an artist
to `new_transform = artist.get_transform() + additional_transform`
with an originally empty (identity) `additional_transform` and then
later adjusts the `additional_transform` by calling `rotate_deg`
or `translate` on it, which has the effect of also modifying
`new_transform` in place.  Instead, directly add the correct
`additional_transform` at the site of use.