Improve ps handling of individual usetex strings.
The ps backend currently determines usetex status solely based on the
rcparam at renderer instantiation time; in particular, having single
Text objects with usetex=True in an otherwise usetex=False renderer
currently results in an AttributeError.  Instead, detect that and do a
best-effort fix which is to try rendering with mathtext instead of
usetex.

(Actually fixing the ps backend would be nicer, but a bit of work.)

Also fix another ps test to not use a temporary directory.