BUG fixed by @anntzer: correct tick labels with subsampled FixedLocator
When explicitly setting the tick locations and then the corresponding
tick labels, a FixedLocator is used for the locations.  It has an nbins
argument that can trigger subsampling.  Previously, tick labels were
made by FixedFormatter and based on position, not value, so subsampling
caused a mismatch, yielding incorrect tick labels.  This is changed to
use FuncFormatter with a function to look up the label by value,
ignoring the position.