Remove randomness from test_colorbar_get_ticks_2.
The point of the test is to check that colorbar ticks have certain
values when the scalarmappable's lower and upper lims are a bit more
than 0 and a bit less than 1, respectively.  Certainly sampling 30x30
random values will nearly always accomplish this, but we may as well
just enforce the values.