Share elevation and azimut with other axes
Just like sharex, sharez et al. enable syncing the view among axes, it should be possible to sync azimut and elevation as well. 
My primary application of this is displaying an interactive stereoscopic (cross-eye) view of 3D plots: 2 Axes3D with the same data are put side-by-side with a difference in azimuth of ~3 degree.