Support odd-length dash patterns in Agg.
All vector backends already support passing an odd-length dash pattern,
duplicating the on/off array (so that on the second repetition "on"
becomes "off" and vice-versa).  This is also explicitly supported by
the svg spec for `stroke-dasharray` ("If an odd number of values is
provided, then the list of values is repeated to yield an even number of
values. Thus, 5,3,2 is equivalent to 5,3,2,5,3,2.") and less explicitly
by the pdf and ps specs, and by cairo.  So the backend needing support
is Agg; iterating twice over the sequence is likely simplest.