FIX: propagate _is_saving state when changing canvases
We have logic to draw animated artists if we are rendering due to
savefig (as opposed to rendering to put into a GUI).  Not every canvas
class can write every format so we can switch the canvas as save time
based on the requested extension / format.  Make sure we set the sate
on the canvas we are actually using to save the figure rather than the
canvas that is attached to the figure when we start saving.

closes #17599