Fix example in docstring of cbook._unfold.
doctest blocks cannot contain empty lines.  In theory they should be
replaced by `<BLANKLINE>` (literally), but given that we don't actually
run doctests, it seems nicer to just delete the empty lines.

We didn't detect the bad docstring because the function is private and
thus its doc not built, but this also confuses vim's autoindenter, so...