FIX: correctly process the tick label size
Passing the size positional caused it to be treated as a fontconfig
pattern.  For the relative sizes without '-' in them this formally
works and stashes the relative size in the family.  We don't actually
resolve that we can _find_ the font so this, while wrong, is not
noticed and when we ask for the size it gives us back the default
size so our size estimate was silently wrong.

In the cases where we have a '-' in the relative size the fontconfig
pattern parsing fails when we try to estimate the size.

closes  #17670

Co-authored-by: Elliott Sales de Andrade <quantum.analyst@gmail.com>