MNT: use get_fillstyle not is_filled to null facecolor
This is brittle, but matches the behavior in Line2D.

MarkerStyle objects have two coupled, but not fully redundant methods
for determining if the maker is filled: the `is_filled` and
`get_fillstyle` methods.  If `ms.get_fillstyle() == 'none'` then
`ms.is_filled() is False`, however the converse is not True.  In
particular the markers that can not be filled (because the Paths they
are made out of can not be closed) have `ms.get_fillstyle() == 'full'`
and `ms.is_filled() is False`.  In Line2D we filter on the value of
`get_fillstyle` not on `is_filled` so do the same in `Axes.scatter`.

In Line2D we do the validation at draw time (because Line2D holds onto
its MarkerStyle object instead of just extracting the path).

The logic for fillstyle on Markers came in via #447/
213459eabb44b4b84bfafc51a9c6964f1e0494cf.

closes #17849

Revises #17543 / d86cc2bab8183fd3288ed474e4dfd33e0f018908

Co-authored-by: Tim Hoffmann <2836374+timhoffm@users.noreply.github.com>
Co-authored-by: Jody Klymak <jklymak@gmail.com>
Co-authored-by: Tim Hoffmann <2836374+timhoffm@users.noreply.github.com>