Force origin='upper' in pyplot.specgram (#17897)
* Force origin='upper' in specgram function

Specify origin='upper' in call to imshow within specgram to override rcparams setting for origin without changing rcparams itself. Delete origin from kwargs so there are not two occurrences of the origin parameter in the call to imshow (this also means the origin kwarg cannot be changed by the caller).

* DOC: indicate that origin is force set to upper

* Adding test for issue 17878

* DOC: Change rcParam from 'origin' to 'image.origin'

* TST: try changing origin in specgram call

* Document behavior change for specgram

* Document api behavior change

* Raise exception if origin kwarg passed to specgram

This ensures the origin kwarg is not silently ignored if passed to pyplot.specgram.

* TST: passing origin kwarg to specgram should fail

The function pyplot.specgram should raise a TypeError if origin is passed as a keyword argument.

* DOC: Remove note that origin='upper' in specgram

This note referenced the call to imshow inside of specgram, which might confuse users who do not know the implementation of specgram.

* TST: change from pytest xfail to pytest raises

Passing origin as a keyword argument should raise a type error, so test_specgram_origin_kwarg should raise an exception. Therefore, raises is a more precise check than xfail.

Co-authored-by: David Stansby <dstansby@gmail.com>
Co-authored-by: Theodor Athanasiadis <theodorathanasiadis@gmail.com>
Co-authored-by: Tim Hoffmann <2836374+timhoffm@users.noreply.github.com>