Always strip out date in postscript's test_savefig_to_stringio.
We already test that postscript output supports SOURCE_DATE_EPOCH in
test_determinism.  Removing the date in test_savefig_to_stringio in all
cases makes it easier to reuse the test in mplcairo, as cairo does not
support SOURCE_DATE_EPOCH (or manually setting the date in postscript
output in any case); the test is not about SOURCE_DATE_EPOCH support
anyways.