Better document multilinebaseline.
Also set its default to False instead of None.  (We only check
truthiness, so False and None are equivalent.)

Also don't explicitly set the default verticalalignment in textprops to
"baseline" -- that is already the normal default of `Text`, and is thus
redundant; moreover, the old implementation would cause property
collision if "verticalalignment" (fully spelt out) was passed in
instead (basically, a call to `normalize_kwargs` was missing).

Also minor misc. docstring improvements.