Make the formatting more explicit
This avoids exponential notation (which is supported by PS but who
knows how all the different parsers react to it) and trailing zeros,
so we don't unnecessarily increase the size of fonts that have a
nice [0.001 0 0 0.001] matrix.

Encapsulate the formatting operation in a cbook function and give
it a unit test.