Revert "Convert adjust_bbox to use ExitStack."
With the `ExitStack` and context managers, it's expected that everything
is called in pairs, i.e., `adjust_bbox` followed by `restore_bbox`. When
saving, `BackendBases.print_figure` calls `adjust_bbox` before and
`restore_bbox` after the renderer-specific method. In raster renderers,
that's all that happens.

However, in mixed mode renderers, this is not the case. When it needs to
rasterize (`MixedModeRendering.start_rasterizing`), it will change the
DPI, restore the bbox, and re-adjust again. Internally, it hangs on to
the new `restore_bbox`, but `print_figure` still has the original, which
was cleared at the first `start_rasterizing` call, so when it calls
`restore_bbox`, it does nothing because the stack is empty.

So all that means that we need to revert the original change, as it
needs to be possible to call `restore_bbox` more than once. This has not
been a direct revert, as other cleanups happened in the meantime.