Inline ScalarFormatter._formatSciNotation.
It's only used in one place, in a tiny function, and doesn't have a
particularly intuitive name, so it's actually easier to follow the logic
by having it all in one place.  Moreover the implementation can be
greatly simplified by separately computing and formatting the mantissa
and the exponent, rather than trying to parse the locale-dependent
result of the formatting of the entire number.