Correctly position text with nonzero descent with afm fonts / ps output.
See e.g.
```
rcParams["ps.useafm"] = True
text(.5, .5, "qk")
axhline(.5)
savefig("/tmp/test.ps")
```
Previously the bottom of the tip of the "q" would be on the baseline;
now the baseline is correctly positioned.

(I picked a relatively generic name for the test because we could later
update the baseline image "in place" for more afm tests.)