Fold ColorbarPatch into Colorbar, deprecate colorbar_factory.
It is easier to have an internal switch on the mappable type, rather
than having a separate "internal" (per the documentation) API to support
hatchings.  Also note that the previous version would fail in obscure
cases where the contourset was changed from not using hatches to using
hatches after the colorbar was created.