Consistent behavior of draw_if_interactive across interactive backends.
Previously it would call `show()` on tk and nbagg, but only
`draw_idle()` on the other backends.  I would guess this was an accident
of history.

Remove that discrepancy, and the customization point
(`trigger_manager_draw`) that I previously implemented to repro
the historical behavior. (`trigger_manager_draw` is in the private
`_Backend` class, so doesn't need a deprecation period.)