Use selectfont instead of findfont + scalefont + setfont in PostScript.
`key scale selectfont` is literally defined as `key findfont scale
scalefont setfont`, but "almost always more efficient" (according to the
spec).

Also unify the AFM and non-AFM codepaths.

I decided to just change the behavior of `set_font` in the AFM case as
it seems unlikely people were relying on that; the alternative would be
to just completely deprecate `set_font` in favor of a new `_set_font`
with the new behavior, which in fact seems more disruptive...