Prepare for merging SubplotBase into AxesBase.
It should be possible to merge SubplotBase into AxesBase, with all Axes
having a `get_subplotspec()` method -- it's just that that method would
return None for non-gridspec-positioned Axes.  The advantage of doing so
is that we could get rid of the rather hackish subplot_class_factory,
and the dynamically generated AxesSubplot class, which appears nowhere
in the docs.

- Deprecate most Subplot-specific API: while it's fine for all axes to
  have a `get_subplotspec` which may return None, it would be a bit
  weird if they all also have e.g. a `is_last_row` for which it's not
  clear what value to return for non-gridspec-positioned Axes.  Moving
  that to the SubplotSpec seems natural enough (and these are pretty
  low-level anyways).

- Make most parameters to AxesBase keyword-only, so that we can later
  overload the positional parameters to be either a rect or a
  subplot triplet (which should ideally be passed packed as a single
  parameter rather than unpacked, but at least during the deprecation
  period it would be a pain to differentiate whether, in `Axes(fig, a,
  b, c)`, `b` was intended to be the second index of a subplot triplet
  or a `facecolor`...)

Due to the order of calls during initialization, Axes3D self-adding to
the figure was problematic.  This is already getting removed in another
PR, so I included the same change here without API changes notes just to
get the tests to pass.  However I can put a note in for this PR if it
ends up being ready for merge first.