Small cleanups to parasite axes.
Setting axis zorder in cla() is unnecessary since 51f8468 (zorders are
now set in set_axisbelow() instead of draw()).

Deprecate the peculiar `update_viewlim` API which is really just an
internal helper.

No need to explicitly set frameon=False, this is forced by the
ParasiteAxesBase constructor.

We can factor the various twin axes removers into a single one.  (Also,
avoiding local functions helps picklability, as usual.)

In twin() we can also use viewlim_mode="transform" when the transform is
None (IdentityTransform()), because that just means that we apply an
additional IdentityTransform() to the viewlims.

In parasite_simple2 we don't need to set viewlim_mode, this is already
done by `twin`.