Fix singleshot timers in wx.
`TimerWx._timer_set_single_shot` currently tries to start the timer
without interval or any other setting, which is not what this method is
supposed to do; it should only modify settings on internal objects.
Since single shot mode is fully handled in `TimerWx._timer_start`, this
method is unnecessary.