FIX: make sure we have more than 1 tick with small log ranges
If we have both a small target number of ticks and a small displayed
range we would previously pick a tick stride that was the same size or
bigger than the visible data range which would result in only 1 tick
being visible on the axis.

This patch ensures that, with a floor at 1, the stride is smaller than
the estimated data range.

re-closes #8768  (same symptoms, different cause)