Make a set_offsets function for the quiver
If you increase the number of arrows with set_offsets and then try to update the (U,V) with set_UVC you will get a ValueError even if both are the same dimension.
The total number of arrows doesn't update as the Quiver class doesn't have a proper set_offsets function.