Support `ax.grid(visible=<bool>)`.
Remove the separate _gridOnMajor/_gridOnMinor attributes and store
visibility together with all other grid attributes in the
_major_tick_kw/_minor_tick_kw dicts under the "gridOn" key.  This makes
it possible to update just that entry in `Axis.grid`.

Also, don't normalize `b` to True `if len(kwargs)` in `Axes.grid`:
we need to keep it as `None`, normalizing it only to True/False in
`Axis.grid` (which handles that just fine), so that `Axis.grid` can
handle the consistency checks between `b` and `visible`.