FIX: allow Manager classes to take string num
The docstring in ManagerBase claims that *num* can be str or int,
however ever backend used int-specific formatting so this was not
actually true.

Fixed the backends rather than fixing the docstring (because setting
the window title on Manager init is useful and implicitly supported
through some special-casing in `plt.figure`).