Cleanup stem docs and simplify implementation.
On the implementation side, we can use `vlines`/`hlines` to generate the
correct `LineCollection`.

On the documentation side, it seems perfectly valid to only specify a
linestyle and no color in `linefmt` (e.g.
`stem([0, 1], [2, 3], linefmt="-.")` works just fine).  On the other
hand markers are indeed silently dropped (possibly we could emit a
warning there, but that should be another PR).  So reword accordingly;
and also slighly reword the doc for `markerfmt`.