MNT: Remove deprecated axes kwargs collision detection
In Matplotlib 2.1, the behavior of reusing existing axes when
created with the same arguments was deprecated (see #9037). This
behavior is now removed.

Functions that create new axes (`axes`, `add_axes`, `subplot`, etc.)
will now always create new axes, regardless of whether the kwargs
passed to them match already existing axes.

Passing kwargs to `gca` is deprecated. If `gca` is called with
kwargs that do not match the current axes, then an exception is
raised.

Fixes #18832.